analytics_project/requirements.txt
pandas>=1.3
matplotlib
seaborn
jupyter
notebook
openpyxl
pyyaml
pytest
numpy
